ai_service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
mysql-connector-python
sentence-transformers
faiss-cpu
numpy
